LibWeb/CSS: Fix implicit narrowing cast in interpolate_rotate
The changes introduced in 484a09d6a2c59c3ffdb903df777913ccbcf7ac79
broke the Windows LibWeb build, so this fixes that

+++ Libraries/LibWeb/CSS/Interpolation.cpp
@@ -224,7 +224,7 @@ static RefPtr<CSSStyleValue const> interpolate_rotate(DOM::Element& element, Cal
     auto from_axis_angle = [](FloatVector3 const& axis, float angle) -> FloatVector4 {
         auto normalized = axis.normalized();
         auto half_angle = angle / 2.0f;
-        auto sin_half_angle = sin(half_angle);
+        auto sin_half_angle = sinf(half_angle);
         FloatVector4 result { normalized.x() * sin_half_angle, normalized.y() * sin_half_angle, normalized.z() * sin_half_angle, cosf(half_angle) };
         return result;
     };
